add events to technology chapter
adds three events to the "technology" chapter of the summary file. the events added are node.js conference brescia 8 2023, rust lab florence 10 2023, and pycon italia florence 4 2024. these events are added to the newly created `events.md` file in the `tech` directory. the `main.rs` file is also updated to fix a minor bug in the build process.